chore: 테스트 클래스 이름 변경 UserLottosTestGame.java -> UserLottosTest.java
